+++ ICD.Connect.Audio.Shure/AbstractShureMxaDeviceSettings.cs
@@ -1,37 +1,6 @@
-﻿using ICD.Common.Utils.Xml;
-using ICD.Connect.Devices;
-using ICD.Connect.Protocol.Ports;
-using ICD.Connect.Settings.Attributes.SettingsProperties;
-
-namespace ICD.Connect.Audio.Shure
+﻿namespace ICD.Connect.Audio.Shure
 {
-	public abstract class AbstractShureMxaDeviceSettings : AbstractDeviceSettings
+	public abstract class AbstractShureMxaDeviceSettings : AbstractShureMicDeviceSettings
 	{
-		private const string PORT_ELEMENT = "Port";
-
-		[OriginatorIdSettingsProperty(typeof(ISerialPort))]
-		public int? Port { get; set; }
-
-		/// <summary>
-		/// Writes property elements to xml.
-		/// </summary>
-		/// <param name="writer"></param>
-		protected override void WriteElements(IcdXmlTextWriter writer)
-		{
-			base.WriteElements(writer);
-
-			writer.WriteElementString(PORT_ELEMENT, IcdXmlConvert.ToString(Port));
-		}
-
-		/// <summary>
-		/// Updates the settings from xml.
-		/// </summary>
-		/// <param name="xml"></param>
-		public override void ParseXml(string xml)
-		{
-			base.ParseXml(xml);
-
-			Port = XmlUtils.TryReadChildElementContentAsInt(xml, PORT_ELEMENT);
-		}
 	}
 }