+++ ICD.Connect.Audio.Shure/ShureMxwApt4DeviceSettings.cs
@@ -1,0 +1,9 @@
+﻿using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Audio.Shure
+{
+	[KrangSettings("ShureMxwApt4", typeof(ShureMxwApt4Device))]
+	public class ShureMxwApt4DeviceSettings : AbstractShureMicDeviceSettings
+	{
+	}
+}